test(import): split ImportAllocationsMessageHandlerTest by run path
Group run, reimport, failure/outcome, and message-bus coverage with a shared Foundry test case.
